frontend/requirements.txt
flask
sockets
requests
pycryptodome
